frontend, reduced to smaller components
