+++ client/src/components/ui/card/CardActivity.jsx
@@ -1,0 +1,107 @@
+import React from "react";
+import ActivityMember from "./ActivityMember";
+// import { useSelector } from "react-redux";
+
+const CardActivity = ( { card }) => {
+  const comments = card.comments;
+  const actions = card.actions;
+
+  console.log(comments, actions);
+  /*
+  Plan for <ActivityMember /> component
+    1. get comments and actions from state
+    2. combine into a allMembers array
+    3. sort allMembers array by 'createdAt' property
+    4. map allMembers array to render each <ActivityMember /> component
+
+  */
+
+  return (
+    <li className="activity-section">
+      <h2 className="activity-icon icon">Activity</h2>
+      <ul className="horiz-list">
+        <li className="not-implemented">Show Details</li>
+      </ul>
+      <ul className="modal-activity-list">
+        <ActivityMember />
+        <li>
+          <div className="member-container">
+            <div className="card-member">TP</div>
+          </div>
+          <h3>Taylor Peat</h3>
+          <div className="comment static-comment">
+            <span>The activities are not functional.</span>
+          </div>
+          <small>
+            22 minutes ago - <span className="link">Edit</span> -{" "}
+            <span className="link">Delete</span>
+          </small>
+          <div className="comment">
+            <label>
+              <textarea required="" defaultValue={'The activities have not been implemented yet.'}rows="1">
+              </textarea>
+              <div>
+                <a className="light-button card-icon sm-icon"></a>
+                <a className="light-button smiley-icon sm-icon"></a>
+                <a className="light-button email-icon sm-icon"></a>
+              </div>
+              <div>
+                <p>You haven&apos;t typed anything!</p>
+                <input
+                  type="submit"
+                  className="button not-implemented"
+                  value="Save"
+                />
+                <i className="x-icon icon"></i>
+              </div>
+            </label>
+          </div>
+        </li>
+        <li>
+          <div className="member-container">
+            <div className="card-member small-size">VR</div>
+          </div>
+          <p>
+            <span className="member-name">Victor Reyes</span> changed the
+            background of this board <small>yesterday at 4:53 PM</small>
+          </p>
+        </li>
+        <li>
+          <div className="member-container">
+            <div className="card-member">VR</div>
+          </div>
+          <h3>Victor Reyes</h3>
+          <div className="comment static-comment">
+            <span>Example of a comment.</span>
+          </div>
+          <small>
+            22 minutes ago - <span className="link">Edit</span> -{" "}
+            <span className="link">Delete</span>
+          </small>
+          <div className="comment">
+            <label>
+              <textarea required="" defaultValue={"Example of a comment."} rows="1">
+              </textarea>
+              <div>
+                <a className="light-button card-icon sm-icon"></a>
+                <a className="light-button smiley-icon sm-icon"></a>
+                <a className="light-button email-icon sm-icon"></a>
+              </div>
+              <div>
+                <p>You haven&apos;t typed anything!</p>
+                <input
+                  type="submit"
+                  className="button not-implemented"
+                  value="Save"
+                />
+                <i className="x-icon icon"></i>
+              </div>
+            </label>
+          </div>
+        </li>
+      </ul>
+    </li>
+  )
+}
+
+export default CardActivity;
