+++ client/src/components/ui/card/Labels.jsx
@@ -1,0 +1,33 @@
+import React from "react";
+
+const Labels = () => {
+
+  return (
+    <li className="labels-section">
+      <h3>Labels</h3>
+      <div className="member-container">
+        <div className="green label colorblindable"></div>
+      </div>
+      <div className="member-container">
+        <div className="yellow label colorblindable"></div>
+      </div>
+      <div className="member-container">
+        <div className="orange label colorblindable"></div>
+      </div>
+      <div className="member-container">
+        <div className="blue label colorblindable"></div>
+      </div>
+      <div className="member-container">
+        <div className="purple label colorblindable"></div>
+      </div>
+      <div className="member-container">
+        <div className="red label colorblindable"></div>
+      </div>
+      <div className="member-container">
+        <i className="plus-icon sm-icon"></i>
+      </div>
+    </li>
+  )
+}
+
+export default Labels;
