+++ client/src/constants/ApiRoutes.js
@@ -5,3 +5,4 @@ export const CREATE_LIST_URL = "/api/lists/";
 export const EDIT_LIST_URL = "/api/lists/";
 export const CREATE_CARD_URL = "/api/cards";
 export const CARD_BY_ID = "/api/cards/";
+export const CREATE_COMMENT_URL = "/api/comments";
